+++ src/pages/Home/Home.tsx
@@ -1,0 +1,147 @@
+import type { NextPage } from 'next'
+import Link from 'next/link';
+import Head from 'next/head';
+
+const Home: NextPage = () => {
+  return (
+    <>
+      <Head>
+        <title>Хакспейс в Санкт-Петербурге • B4CKSP4CE</title>
+      </Head>
+
+      <h1>B4CKSP4CE</h1>
+      <article>
+        <p>
+          Бэкспейс — это <a href="https://ru.wikipedia.org/wiki/Хакерспейс">хакспейс</a>.
+          Сообщество энтузиастов и энтузиасток, обьединённых вокруг технологий и искусства.
+          Вместе мы разбираемся в сложных вещах, проводим события, работаем над проектами и 
+          просто хорошо проводим время.
+        </p>
+        <p>
+          Наша миссия – развивать сообщество, в котором все смогут найти что-то для себя,
+          Мы ломаем стены и строим мосты, помогая друг другу и сообществам вокруг нас.
+        </p>
+      </article>
+
+      <article>
+        <header>
+          <h2>Как с нами взаимодействовать</h2>
+        </header>
+        <p>
+          Приходите к нам на события – мы проводим мастер-классы, лекции и встречи обо всем –
+          от пайки и пентеста, до синтезаторов, шейдеров и VR.
+        </p>
+        <p>
+          Приходите в гости – пообщаться, познакомиться с пространством или попросить помощи
+          для своего проекта. Мы понимаем, что всем комфортно знакомиться по-разному, поэтому
+          можно прийти просто почитать книгу (у нас классная библиотека) или попросить резидента
+          помочь познакомиться с остальными.
+        </p>
+        <p>
+          Резиденты занимаются Бэкспейсом в свободное время на чистом энтузиазме, поэтому у нас нет
+          “дежурных” или “администраторов”. Мы очень любим гостей, но встречаем их только когда кто-то
+          из резидентов готов это сделать. Собираясь заглянуть, напишите заранее в&nbsp;
+          <a href="https://t.me/b4cksp4ce_chat">чат</a>,
+          чтобы мы согласовали удобное всем время.
+        </p>
+        <p>
+          Бэкспейс – не закрытый клуб. Если вы разделяете наши принципы и хотите делать его вместе – станьте <a href="https://wiki.0x08.in/ru/101/roles">резидентом</a>.
+        </p>
+      </article>
+
+      <article>
+        <header>
+          <h2>Принципы</h2>
+        </header>
+        <p>Наше сообщество построено на нескольких принципах:</p>
+        <ol>
+          <li>
+            <b>Будьте прекрасны по отношению друг к другу</b>
+            <p>Прислушивайтесь ко мнению и потребностям друг друга.</p>
+          </li>
+          <li>
+            <b>Не притесняй и не мешай другим</b>
+            <p>Соблюдай личные границы других участников, старайся договариваться и кооперироваться.</p>
+          </li>
+          <li>
+            <b>Вкладывай больше, чем потребляешь</b>
+            <p>
+              Бэкспейс может развиваться только если сообщество делает его лучше.
+              Мы не коворкинг и не обеспечиваем резидентов готовыми условиями для работы.
+            </p>
+          </li>
+          <li>
+            <b>Финансовая независимость</b>
+            <p>Бэкспейс не является источником прибыли, а финансовый вклад резидента не влияет на принятие решений.</p>
+          </li>
+          <li>
+            <b>Do-ocracy</b>
+            <p>Если хочешь чтобы что-то изменилось – сделай это, но не забывай о других.</p>
+          </li>
+          <li>
+            <b>Запрещено умирать</b>
+            <p>Safety first. Заботьтесь о безопасности себя и окружающих.</p>
+          </li>
+        </ol>
+        <p>Подробнее, в <a href="https://wiki.0x08.in/ru/101/rules-and-principles">B4CKSP4CE 101</a>. </p>
+      </article>
+
+      <article>
+        <header>
+          <h2>Донейшн</h2>
+        </header>
+        <p>
+          Бэкспейс – это горизонтальное сообщество, существующее только благодаря резидентским взносам и пожертвованиям.
+          Если вы хотите нас поддержать, можно:
+        </p>
+        <ul>
+          <li>задонатить на <a href="https://sobe.ru/na/bksp">Собе.ру</a></li>
+          <li>закинуть наличку в Donation Box</li>
+          <li>
+            задонатить что-нибудь из&nbsp;
+            <Link href="/wishlist" passHref>
+              <a>вишлиста</a>
+            </Link>
+          </li>
+        </ul>
+      </article>
+
+      <article>
+        <header>
+          <h2>Контакты</h2>
+        </header>
+        <ul>
+          <li>
+            СПб, Ланское шоссе, 14к1 (м. Черная речка) —&nbsp;
+            <a href="https://yandex.ru/maps/-/CCU5yWBvGD">Яндекс</a>,&nbsp;
+            <a href="https://go.2gis.com/nf715">2GIS</a>
+          </li>
+          <li>
+            Подробнее о том,&nbsp;
+            <Link href="/where" passHref>
+              <a>как до нас добраться</a>
+            </Link>&nbsp;
+          </li>
+          <li>
+            <span>Telegram: </span>
+            <a href="https://t.me/b4cksp4ce_l1ve">блог</a>,&nbsp;
+            <a href="https://t.me/b4ck5p4c3">анонсы</a>,&nbsp;
+            <a href="https://t.me/b4cksp4ce_chat">чат</a>,&nbsp;
+            <a href="https://t.me/b4cksp4ce_issues">задачи</a>
+          </li>
+          <li>
+            <a href="https://github.com/b4ck5p4c3">Github</a>,&nbsp;
+            <a href="https://matrix.to/#/#b4ck5p4c3:matrix.org">Matrix</a>
+          </li>
+          <li>
+            <a href="https://www.instagram.com/b4ck5p4ce">Instagram</a>,&nbsp;
+            <a href="https://twitter.com/b4cksp4ce">Twitter</a>
+          </li>
+          <li><a href="tel:78125079241">+7 (812) 507-92-41</a> на случай важных переговоров</li>
+        </ul>
+      </article>
+    </>
+  )
+}
+
+export default Home;
