+++ src/pages/whishlist.tsx
@@ -1,0 +1,2 @@
+import Whishlist from "./Whishlist/Wishlist";
+export default Whishlist;
