+++ src/styles/vars.scss
@@ -1,0 +1,1 @@
+$break-mobile: 768px;
